Move function from DOC to STET
Only STET uses it
